# legion/operator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a EDI server.",
        "title": "EDI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/v1/model/token": {
            "post": {
                "description": "Create a JWT token for access to the model service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "Create a model JWT token",
                "parameters": [
                    {
                        "description": "Create a model JWT token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/v1.TokenResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/routes.HTTPResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "routes.HTTPResult": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success of error message",
                    "type": "string"
                }
            }
        },
...
        "v1.TokenRequest": {
            "type": "object",
            "properties": {
                "expiration_date": {
                    "description": "Explicitly set expiration date for token",
                    "type": "string"
                },
                "model_id": {
                    "description": "Model id",
                    "type": "string"
                },
                "model_version": {
                    "description": "Model version",
                    "type": "string"
                }
            }
        },
        "v1.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}